fetch/metadata/generated/element-frame.sub.tentative.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/element-frame.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>HTTP headers on request for HTML "frame" element source</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(url, test, userActivated) {
    const frame = document.createElement('frame');

    const setSrc = () => frame.setAttribute('src', url);

    document.body.appendChild(frame);
    test.add_cleanup(() => frame.remove());

    return new Promise((resolve) => {
       if (userActivated) {
         test_driver.bless('enable user activation', setSrc);
       } else {
         setSrc();
       }

        frame.onload = frame.onerror = resolve;
      });
  }

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpOrigin'], {mime: 'text/html'}),
        t,
        false
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
        });
  }, 'sec-fetch-frame-ancestors - Not sent to non-trustworthy same-origin destination');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpSameSite'], {mime: 'text/html'}),
        t,
        false
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
        });
  }, 'sec-fetch-frame-ancestors - Not sent to non-trustworthy same-site destination');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpCrossSite'], {mime: 'text/html'}),
        t,
        false
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-frame-ancestors');
        });
  }, 'sec-fetch-frame-ancestors - Not sent to non-trustworthy cross-site destination');
  </script>
  </body>
</html>
